04/05/2022/23/23

+++ ImageCollection/Models/Collection.cs
@@ -1,13 +1,16 @@
 ﻿using ImageCollection.Interfaces;
+using Prism.Mvvm;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
+using System.Windows;
+using SUID = Sergey.UI.Extension.Dialogs;
 
 namespace ImageCollection.Models
 {
-    internal class Collection : ICollection
+    internal class Collection : BindableBase, ICollection
     {
         public Guid Id { get; }
         public string Name { get; set; }
@@ -35,12 +38,20 @@ namespace ImageCollection.Models
 
         public void RemoveSelectedFiles()
         {
-            IEnumerable<ICollectionItem> selectedItems = Items.Where(item => item.IsSelected);
-            string collectionDirectory = Path.Combine(_collectionsManager.RootDirectory, Name);
-            foreach (ICollectionItem collectionItem in selectedItems)
+            IReadOnlyList<ICollectionItem> selectedItems = Items.Where(item => item.IsSelected).ToList();
+            string message = "Удалить выбранные файлы?";
+            if (selectedItems.Count == 1)
             {
-                File.Delete(Path.Combine(collectionDirectory, collectionItem.Name));
-                Items.Remove(collectionItem);
+                message = $"Удалить {selectedItems[0].Name}?";
+            }
+            if (SUID.MessageBox.Show(message, App.Name, MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes)
+            {
+                string collectionDirectory = Path.Combine(_collectionsManager.RootDirectory, Name);
+                foreach (ICollectionItem collectionItem in selectedItems)
+                {
+                    File.Delete(Path.Combine(collectionDirectory, collectionItem.Name));
+                    Items.Remove(collectionItem);
+                }
             }
         }
 
@@ -53,5 +64,10 @@ namespace ImageCollection.Models
         {
             throw new NotImplementedException();
         }
+
+        public override bool Equals(object obj) => (obj is Collection).Equals(Id);
+
+        public override int GetHashCode() => Id.GetHashCode();
+
     }
 }
